avoid breaking a public interface
